refactoring and add metro.config.js
